raftstore: extract coprocessor::properties
